feat: require explicit secret path & support any secret engine
This allows to use different secret mount points in the same configuration,
as well as support engines other than KV v2.
With it also comes the possibility to load an entire secret at once, by
not passing a `vault_secret_key`.

BREAKING CHANGE: This requires that you explicitely pass the full secret
path, eg. `secret/data/<path/to/secret>` for a KV v2, or
`database/creds/<your DB role>` for a database engine.